requirements.txt
streamlit==1.33.0
python-docx==0.8.11
reportlab==4.0.8
pdfplumber==0.10.2
langchain==0.1.14
openai==1.15.0
numpy==1.26.4
faiss-cpu==1.7.4